+++ frontend/src/utils/index.js
@@ -1,0 +1,3 @@
+export function dataSlicer(data, itemsPerPage) {
+  return Math.ceil(data?.length / itemsPerPage);
+}
